helm: removing elasticsearch
Elasticsearch should not be a part of the standard helm chart
deployment since it is not likely that potential users are
going to use a "default" elastic deployment on top of Kubernetes.

Also it would increase the scope of the helm chart, that results
in more maintenance work around it.

Let the users bring their own data store and enable the helm chart
to be configured with data store endpoint.

In the documentation of the helm chart it can be mentioned how to deploy
haystack with elasticsearch that is deployed through the official helm
chart.

Signed-off-by: Greg Nagy <greg@gatecastle.com>