restructure: split UI out of deployment and service
Removing haystack UI out of the deployment and the service the reason
behind this is that generally it is not a good practice to put UI and
API service to the same pod. You don't necessarily want to scale the UI
together with the API service.

On the other hand UI should be an optional deployment (in an other helm
chart) since in production grade setups you usually want to separate the
two. (access, capacity planning reasons)

This commit removes the ui from the helm chart.

Signed-off-by: Greg Nagy <greg@gatecastle.com>